% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{annotate.gsp}
\alias{annotate.gsp}
\alias{annotate.gsp,Phenotypes-method}
\alias{annotate.gsp,gespeR-method}
\title{annotate.gsp}
\usage{
\S4method{annotate.gsp}{Phenotypes}(object, organism = "hsapiens")

\S4method{annotate.gsp}{gespeR}(object, organism = "hsapiens")
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} or \code{\linkS4class{Phenotypes}} object}

\item{organism}{String indicating the biomaRt organism, e.g. "hsapiens"}

\item{...}{Additional arguments}
}
\value{
data.frame containing gene identifier, gene symbol and phenotypic score
}
\description{
Query Biomart HGNC symbols for the entrez identifiers of estimated GSPs.
Currently, only implemented for species "hsapiens".
}
\author{
Fabian Schmich
}

