% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Phenotypes-class.R
\docType{methods}
\name{as.data.frame,Phenotypes-method}
\alias{as.data.frame,Phenotypes-method}
\title{Convert Phenotypes object to a data.frame}
\usage{
\S4method{as.data.frame}{Phenotypes}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}
}
\value{
A data.frame
}
\description{
Convert Phenotypes object to a data.frame
}
\examples{
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
as.data.frame(phenos)
}
\author{
Fabian Schmich
}

