% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gespeR-concordance.R
\name{as.data.frame.concordance}
\alias{as.data.frame.concordance}
\title{Coerce method}
\usage{
\method{as.data.frame}{concordance}(x, ...)
}
\arguments{
\item{x}{concordance object}

\item{...}{additional arguments}
}
\value{
data.frame
}
\description{
Coerce method
}
\author{
Fabian Schmich
}

