% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Phenotypes-class.R
\docType{methods}
\name{c,Phenotypes-method}
\alias{c,Phenotypes-method}
\title{Concatenate Phenotypes objects}
\usage{
\S4method{c}{Phenotypes}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{...}{additional \code{\linkS4class{Phenotypes}} objects}

\item{recursive}{recursive}
}
\value{
A concatenated \code{\linkS4class{Phenotypes}} object
}
\description{
Concatenate Phenotypes objects
}
\examples{
phenos.a <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
phenos.b <- Phenotypes(system.file("extdata", "Phenotypes_screen_B.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
c(phenos.a, phenos.b)
}
\author{
Fabian Schmich
}

