% Generated by roxygen2 (4.0.1): do not edit by hand
\name{concordance}
\alias{concordance}
\title{Evaluate the concordance between Phenotype objects}
\usage{
concordance(..., min.overlap = 1, cor.method = "spearman",
  cor.use = "pairwise.complete.obs", rbo.p = 0.95, rbo.k = NULL,
  rbo.mid = NULL)
}
\arguments{
\item{...}{The phenotypes to be evaluated for concordance}

\item{min.overlap}{The minimum number of overlapping genes required}

\item{cor.method}{A character string indicating which correlation coefficient is to be computed}

\item{cor.use}{A character string indicating a method for computing correlation in the presence of missing values}

\item{rbo.p}{The weighting parameter for rank biased overlap (rbo) in [0, 1]. High p implies strong emphasis on top ranked elements}

\item{rbo.k}{The evaluation depth for rank biased overlap extrapolation}

\item{rbo.mid}{The mid point to split a ranked list, e.g. in order to split positive and negative scores choose mid=0}
}
\value{
A \code{\link{concordance}} object with the following elements:
\item{pair.test}{Indicator of compared phenotypes}
\item{cor}{The correlation between pairs of phenotypes for the same gene}
\item{rbo.top}{The rank biased overlap of genes evaluated at the top of the ranked list}
\item{rbo.bottom}{The rank biased overlap of genes evaluated at the bottom of the ranked list}
\item{jaccard}{The Jaccard index of selected genes}
}
\description{
Measures include the correlation (rho) between pairs of phenotypes for the same gene, the rank biased overlap (\code{\link{rbo}})
of the top and bottom of ranked lists, and the Jaccard index (J) of selected genes.
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{Phenotype}}

\code{\link{plot.concordance}}

\code{\link{rbo}}
}

