% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-generics.R
\docType{methods}
\name{dim,Phenotypes-method}
\alias{dim,Phenotypes-method}
\title{Dimension of a \code{\linkS4class{Phenotypes}} object}
\usage{
\S4method{dim}{Phenotypes}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Phenotypes}} object}
}
\value{
Dimension of the \code{\linkS4class{Phenotypes}} object
}
\description{
Dimension of a \code{\linkS4class{Phenotypes}} object
}
\author{
Fabian Schmich
}

