% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Extract}
\alias{Extract}
\title{Subsetting for Phenotype objects.}
\usage{
\S4method{[}{Phenotypes,ANY,ANY,ANY}(x, i)

\S4method{[}{TargetRelations,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{i}{The subsetting indices}

\item{x}{A \code{\linkS4class{TargetRelations}} object}

\item{i}{The row subsetting indices (siRNAs)}

\item{j}{The column subsetting indeces (genes)}
}
\value{
A \code{\linkS4class{Phenotypes}} object

A \code{\linkS4class{TargetRelations}} object
}
\description{
Subsetting for Phenotype objects.

Subsetting for TargetRelations objects.
}
\author{
Fabian Schmich

Fabian Schmich
}

