% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{method}
\name{[}
\alias{[}
\alias{[,Extract-Phenotypes-methods}
\title{Subsetting for Phenotype objects.}
\usage{
\S4method{[}{Phenotypes,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{i}{The subsetting indices}
}
\value{
A \code{\linkS4class{Phenotypes}} object
}
\description{
Subsetting for Phenotype objects.
}
\author{
Fabian Schmich
}

