% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{method}
\name{[}
\alias{[}
\alias{[,Extract-TargetRelations-methods}
\title{Subsetting for TargetRelations objects.}
\usage{
\S4method{[}{TargetRelations,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{A \code{\linkS4class{TargetRelations}} object}

\item{i}{The row subsetting indices (siRNAs)}

\item{j}{The column subsetting indeces (genes)}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Subsetting for TargetRelations objects.
}
\author{
Fabian Schmich
}

