% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gespeR-class}
\alias{gespeR}
\alias{gespeR,Phenotypes,TargetRelations-method}
\alias{gespeR,numeric,Matrix-method}
\alias{gespeR-class}
\title{gespeR}
\usage{
gespeR(phenotypes, target.relations, ...)

\S4method{gespeR}{Phenotypes,TargetRelations}(phenotypes, target.relations,
  mode = c("cv", "stability"), alpha = 0.5, nbootstrap = 100,
  fraction = 0.67, threshold = 0.9, EV = 1, weakness = 0.8,
  ncores = 1, ...)

\S4method{gespeR}{numeric,Matrix}(phenotypes, target.relations, ...)
}
\arguments{
\item{phenotypes}{The siRNA-spefic phenotypes}

\item{target.relations}{The siRNA-to-gene target relations}

\item{...}{Additional arguments}

\item{mode}{The mode of covariate selectino ("cv" or "stability")}

\item{alpha}{The \code{\link{glmnet}} mixing parameter}

\item{nbootstrap}{The number of bootstrap samples}

\item{fraction}{The fraction for each bootstrap sample}

\item{threshold}{The selection threshold}

\item{EV}{The expected value of wrongly selected elements}

\item{weakness}{The weakness parameter for randomised lasso}

\item{ncores}{The number of cores for parallel computation}
}
\value{
A \code{\linkS4class{gespeR}} object
}
\description{
Class that represents a gespeR model. It contains a SSP \code{\linkS4class{Phenotypes}} and \code{\linkS4class{TargetRelations}}
representing a siRNA knockdown experiment. When the model is fitted, it additionaly contains estimated GSP \code{\linkS4class{Phenotypes}}.
}
\section{Slots}{

\describe{
\item{\code{SSP}}{The observed siRNA-specific phenotypes}

\item{\code{GSP}}{The deconvoluted gene-specific phenotypes}

\item{\code{target.relations}}{The siRNA-to-gene target relations, e.g. predicted by TargetScan}

\item{\code{is.fitted}}{An indicator wheter the gespeR model was fitted}

\item{\code{model}}{The fitted regularized linear regression model}
}}
\examples{
# Read phenotypes
phenos <- lapply(LETTERS[1:4], function(x) {
  sprintf("Phenotypes_screen_\%s.txt", x)
  })
phenos <- lapply(phenos, function(x) {
  Phenotypes(system.file("extdata", x, package="gespeR"),
             type = "SSP",
             col.id = 1,
             col.score = 2)
})  
phenos
plot(phenos[[1]])

# Read target relations
tr <- lapply(LETTERS[1:4], function(x) {
  sprintf("TR_screen_\%s.rds", x)
})
tr <- lapply(tr, function(x) {
  TargetRelations(system.file("extdata", x, package="gespeR"))
})
tr[[1]]
tempfile <- paste(tempfile(pattern = "file", tmpdir = tempdir()), ".rds", sep="")
tr[[1]] <- unloadValues(tr[[1]], writeValues = TRUE, path = tempfile)
tr[[1]]
tr[[1]] <- loadValues(tr[[1]])
tr[[1]]

# Fit gespeR models with cross validation
res.cv <- lapply(1:length(phenos), function(i) {
  gespeR(phenotypes = phenos[[i]],
         target.relations = tr[[i]],
         mode = "cv",
         alpha = 0.5,
         ncores = 1)
})
summary(res.cv[[1]])
res.cv[[1]]
plot(res.cv[[1]])

# Extract scores
ssp(res.cv[[1]])
gsp(res.cv[[1]])
head(scores(res.cv[[1]]))

# Fit gespeR models with stability selection
res.stab <- lapply(1:length(phenos), function(i) {
  gespeR(phenotypes = phenos[[i]],
         target.relations = tr[[i]],
         mode = "stability",
         nbootstrap = 100,
         fraction = 0.67,
         threshold = 0.75,
         EV = 1,
         weakness = 0.8,
         ncores = 1)
})
summary(res.stab[[1]])
res.stab[[1]]
plot(res.stab[[1]])

# Extract scores
ssp(res.stab[[1]])
gsp(res.stab[[1]])
head(scores(res.stab[[1]]))

# Compare concordance between stability selected GSPs and SSPs
conc.gsp <- concordance(lapply(res.stab, gsp))
conc.ssp <- concordance(lapply(res.stab, ssp))

pl.gsp <- plot(conc.gsp) + ggtitle("GSPs\\n")
pl.ssp <- plot(conc.ssp) + ggtitle("SSPs\\n")

if (require(grid)) {
  grid.newpage()
  pushViewport(viewport(layout = grid.layout(1, 2) ) )
  print(pl.gsp, vp = viewport(layout.pos.row = 1, layout.pos.col = 1))
  print(pl.ssp, vp = viewport(layout.pos.row = 1, layout.pos.col = 2))
} else {
  plot(pl.gsp)
  plot(pl.ssp)
}
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{gespeR-package}}

\code{\link{plot.gespeR}}

\code{\link{gsp}}

\code{\link{ssp}}

\code{\link{scores}}

\code{\link{stability}}

\code{\link{target.relations}}
}

