% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gespeR}
\alias{gespeR}
\alias{gespeR,Phenotypes,TargetRelations-method}
\alias{gespeR,numeric,Matrix-method}
\title{Deconvolute off-target confounded RNAi knockdown phenotypes.}
\usage{
gespeR(phenotypes, target.relations, ...)

\S4method{gespeR}{Phenotypes,TargetRelations}(phenotypes, target.relations,
  mode = c("cv", "stability"), ...)

\S4method{gespeR}{numeric,Matrix}(phenotypes, target.relations, mode = c("cv",
  "stability"), ...)
}
\arguments{
\item{phenotypes}{The siRNA-spefic phenotypes}

\item{target.relations}{The siRNA-to-gene target relations}

\item{mode}{The mode of covariate selectino ("cv" or "stability")}

\item{...}{Additional arguments}
}
\value{
A \code{\linkS4class{gespeR}} object
}
\description{
This generic function can handle different types of inputs for phenotype data and target relation matrices. It either reads from two .rds files,
uses a data.frame and matrix, respectively, or re-fits the model of a \code{\linkS4class{gespeR}} object. The actual model is a regularised linear
regression model using an elastic net penalty: \eqn{X = X\beta + \epsilon}. The model can be tuned using the regularisation parameters .. (text see deepSNV!)
}
\author{
Fabian Schmich
}

