% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{gespeR-package}
\alias{gespeR-package}
\title{Package: Gene-Specific Phenotype EstimatoR}
\description{
This package provides a model to deconvolute off-target confounded RNAi knockdown phentypes, and
methods to investigate concordance between ranked lists of (estimated) phenotypes. The regularized
linear regression model can be fitted using two different strategies. (a) Cross-validation over
regularization parameters optimising the mean-squared-error of the model and (b) stability selection
of covariates (genes) based on a method by Nicolai Meinshausen et al.
}
\examples{
print("Hello. This is is an example.")
}
\author{
Fabian Schmich | Computational Biology Group, ETH ZURICH | \email{fabian.schmich@bsse.ethz.ch}
}
\references{
Fabian Schmich et. al,  Deconvoluting Off-Target Confounded RNA Interference Screens (2014).
}
\seealso{
\code{\link{gespeR}}
}
\keyword{package}

