% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gsp}
\alias{gsp}
\alias{gsp,gespeR-method}
\title{Get GSPs}
\usage{
gsp(object)

\S4method{gsp}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object of GSPs
}
\description{
Get GSPs
}
\author{
Fabian Schmich
}

