% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-methods.R
\docType{methods}
\name{gsp}
\alias{gsp}
\alias{gsp,gespeR-method}
\alias{ssp}
\alias{ssp,gespeR-method}
\title{Retrieve GSPs and SSPs from \code{\linkS4class{gespeR}} objects}
\usage{
gsp(object)

\S4method{gsp}{gespeR}(object)

ssp(object)

\S4method{ssp}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object of GSPs and SSPs, respectively
}
\description{
Retrieve GSPs and SSPs from \code{\linkS4class{gespeR}} objects
}
\examples{
data(stabilityfits)
gsp(stabilityfits$A)
ssp(stabilityfits$B)
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{annotate.gsp}}

\code{\link{scores}}
}

