% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lasso.rand}
\alias{lasso.rand}
\title{Randomized Lasso}
\usage{
lasso.rand(x, y, weakness = 1, subsample = 1:nrow(x), dfmax = (ncol(x) +
  1), lambda = NULL, standardize = FALSE, intercept = FALSE, ...)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{weakness}{The weakness parameter}

\item{subsample}{The data subsample (default: none)}

\item{dfmax}{The maxiumum number of degrees of freedom}

\item{lambda}{The regularisation parameter}

\item{standardize}{Indicator, wheter to standardize the design matrix}

\item{intercept}{Indicator, whether to fit an intercept}

\item{...}{Additional arguments to \code{\link{glmnet}}}
}
\value{
A \code{\link{glmnet}} object
}
\description{
Based on Meinshausen and Buehlmann (2009)
}
\examples{
y <- rnorm(50)
 x <- matrix(runif(50 * 20), ncol = 20)
 lasso.rand(x = x, y = y)
}
\author{
Fabian Schmich
}

