% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{loadMatrix}
\alias{loadMatrix}
\alias{loadMatrix,TargetRelations-method}
\title{Load matrix of a TargetRelations object}
\usage{
loadMatrix(object)

\S4method{loadMatrix}{TargetRelations}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Load matrix of a TargetRelations object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{unloadMatrix}}
}

