% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{loadValues}
\alias{loadValues}
\alias{loadValues,TargetRelations-method}
\alias{loadValues,gespeR-method}
\title{Load values of a TargetRelations object}
\usage{
loadValues(object)

\S4method{loadValues}{TargetRelations}(object)

\S4method{loadValues}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Load values of a TargetRelations object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{unloadValues}}
}

