% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{na.rem}
\alias{na.rem}
\alias{na.rem,Phenotypes-method}
\title{Remove NA/Inf values from phenotype vectors}
\usage{
na.rem(object)

\S4method{na.rem}{Phenotypes}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{Phenotypes}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object without NA scores values
}
\description{
Remove NA/Inf values from phenotype vectors
}
\author{
Fabian Schmich
}

