% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{path<-}
\alias{path<-}
\alias{path<-,TargetRelations,character-method}
\title{Set the path of a A \code{\linkS4class{TargetRelations}} object object}
\usage{
path(object) <- value

\S4method{path}{TargetRelations,character}(object) <- value
}
\arguments{
\item{object}{A \code{\link{TargetRelations}} object}

\item{value}{A string defining the path}
}
\value{
A \code{\linkS4class{TargetRelations}} object with set path
}
\description{
Set the path of a A \code{\linkS4class{TargetRelations}} object object
}
\author{
Fabian Schmich
}

