% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Phenotypes-class.R
\docType{class}
\name{Phenotypes-class}
\alias{Phenotypes}
\alias{Phenotypes,Matrix-method}
\alias{Phenotypes,cellHTS-method}
\alias{Phenotypes,character-method}
\alias{Phenotypes-class}
\title{Phenotypes}
\usage{
Phenotypes(phenotypes, ...)

\S4method{Phenotypes}{character}(phenotypes, type = c("SSP", "GSP"),
  sep = "\\t", col.id = 1, col.score = 2)

\S4method{Phenotypes}{cellHTS}(phenotypes, channel, sample)

\S4method{Phenotypes}{Matrix}(phenotypes, ids = NULL, pnames = NULL,
  type = c("SSP", "GSP"))
}
\arguments{
\item{phenotypes}{The phenotypes as numeric vector, path to a .txt file with two columns (1: identifiers, 2: values), or a cellHTS object}

\item{...}{Additional arguments}

\item{type}{The type of phenotype (GSP, SSP)}

\item{sep}{The separator string}

\item{col.id}{Column number for the siRNA identifiers}

\item{col.score}{Column number(s) for the phenotype score}

\item{channel}{The cellHTS channel identifier}

\item{sample}{The cellHTS sample index}

\item{ids}{The siRNA/gene identifiers}

\item{pnames}{The phenotype identifiers}
}
\value{
A \code{\linkS4class{Phenotypes}} object
}
\description{
Class used to represent various types of phenotypes, e.g. from
siRNA-specific (SSP) or estimated gene-specific phenotypes (GSP).
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of represented phenotypes (i.e., "SSP" or "GSP")}

\item{\code{ids}}{The entity identifiers (i.e., siRNA or gene ids)}

\item{\code{pnames}}{The phenotype names}

\item{\code{values}}{The phenotypic values}
}}
\examples{
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{plot.Phenotypes}}

\code{\link{join}}

\code{\link{gsp}}

\code{\link{ssp}}

\code{\link{scores}}

\code{\link{concordance}}
}

