% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Phenotypes}
\alias{Phenotypes}
\alias{Phenotypes,cellHTS-method}
\alias{Phenotypes,character-method}
\alias{Phenotypes,numeric-method}
\title{Constructor for Phenotypes objects}
\usage{
Phenotypes(phenotypes, ...)

\S4method{Phenotypes}{character}(phenotypes, type = c("SSP", "GSP"),
  sep = "\\t", col.id = 1, col.score = 2)

\S4method{Phenotypes}{cellHTS}(phenotypes, channel, sample)

\S4method{Phenotypes}{numeric}(phenotypes, ids = NULL, type = c("SSP",
  "GSP"))
}
\arguments{
\item{phenotypes}{The phenotypes as numeric vector, path to a .txt file with two columns (1: identifiers, 2: values), or a cellHTS object}

\item{ids}{The phenotype identifiers}

\item{type}{The type of phenotype (GSP, SSP)}

\item{...}{Additional arguments}

\item{sep}{The separator string}

\item{col.id}{Column number for the identifier}

\item{col.score}{Column number for the phenotype score}

\item{channel}{The cellHTS channel identifier}

\item{sample}{The cellHTS sample index}
}
\value{
A \code{\link{Phenotypes}} object
}
\description{
This generic function can handle different types of inputs for phenotype data. It either takes a numeric vector or reads the values from a .txt file.
}
\author{
Fabian Schmich
}

