% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.concordance}
\alias{plot.concordance}
\title{Plot concordance}
\usage{
\method{plot}{concordance}(x)
}
\arguments{
\item{x}{The data of class \code{\link{concordance}}}
}
\description{
Plots boxplots of concordance evaluated between multiple Phenotype objects. Measures include the correlation (rho) between pairs of
phenotypes for the same gene, the rank biased overlap (rbo) of the top and bottom of ranked lists, and the
Jaccard index (J) of selected genes.
}
\author{
Fabian Schmich
}

