% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.gespeR}
\alias{plot.gespeR}
\title{Plot method for \code{\linkS4class{gespeR}} objects}
\usage{
\method{plot}{gespeR}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{gespeR}} object}
}
\description{
Plot method for \code{\linkS4class{gespeR}} objects
}
\author{
Fabian Schmich
}

