% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{scores}
\alias{scores}
\alias{scores,Phenotypes-method}
\alias{scores,gespeR-method}
\title{scores}
\usage{
\S4method{scores}{Phenotypes}(object)

\S4method{scores}{gespeR}(object, type)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} or \code{\linkS4class{Phenotypes}} object}

\item{type}{The type of phenotype scores (GSP, SSP)}
}
\value{
A named vector of scores for each phenotype identifier
}
\description{
Return a named vector of phenotype scores
}
\author{
Fabian Schmich
}
\seealso{
\code{\linkS4class{gespeR}}, \code{\linkS4class{Phenotypes}}
}

