% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,gespeR-method}
\alias{gespeR-method}
\alias{show,}
\alias{show,gespeR-method}
\title{Show method for gespeR objects}
\usage{
\S4method{show}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\description{
Show method for gespeR objects
}
\author{
Fabian Schmich
}

