% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,Phenotypes-method}
\alias{show,Phenotypes-method}
\title{Show method for Phenotype objects}
\usage{
\S4method{show}{Phenotypes}(object)
}
\arguments{
\item{object}{A \code{\link{Phenotype}} object.}
}
\description{
Show method for Phenotype objects
}
\author{
Fabian Schmich
}

