% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ssp}
\alias{ssp}
\alias{ssp,gespeR-method}
\title{Get SSPs}
\usage{
ssp(object)

\S4method{ssp}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object of SSPs
}
\description{
Get SSPs
}
\author{
Fabian Schmich
}

