% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-functions.R
\name{stability.selection}
\alias{stability.selection}
\title{Stability Selection}
\usage{
stability.selection(x, y, fraction = 0.5, threshold = 0.75, EV = 1,
  nbootstrap = 100, weakness = 1, intercept = FALSE, ncores = 1, ...)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{fraction}{The fraction for each bootstrap sample}

\item{threshold}{The selection threshold}

\item{EV}{The expected value of wrongly selected elements}

\item{nbootstrap}{The number of bootstrap samples}

\item{weakness}{The weakness parameter for randomised lasso}

\item{intercept}{Indicator, whether to fit an intercept}

\item{ncores}{The number of cores for parallel computation}

\item{...}{Additional arguments to \code{\link{lasso.rand}}}
}
\value{
A list containing selected covariates with frequencies, and the fitted model
}
\description{
Based on Meinshausen and Buehlmann (2009)
}
\author{
Fabian Schmich
}

