% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{stabilityfits}
\alias{stabilityfits}
\title{Example fits for phenotypes from example screening data A, B, C and D}
\description{
The data set contains four fitted gespeR models using stability selection
}
\examples{
data(stabilityfits)
}

