% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,gespeR-method}
\alias{summary,gespeR-method}
\title{Summary method for \code{\linkS4class{gespeR}} objects}
\usage{
\S4method{summary}{gespeR}(object)
}
\arguments{
\item{object}{an object for which a summary is desired.}
}
\description{
Summary method for \code{\linkS4class{gespeR}} objects
}
\author{
Fabian Schmich
}

