% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{target.relations}
\alias{target.relations}
\alias{target.relations,gespeR-method}
\title{target.relations}
\usage{
target.relations(object)

\S4method{target.relations}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Retrieve siRNA-to-gene target relations from a \code{\linkS4class{gespeR}} object.
}
\author{
Fabian Schmich
}

