% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{TargetRelations}
\alias{TargetRelations}
\title{Class definition for TargetRelations}
\description{
Class used to represent siRNA-to-gene on- and off-target relations
for a knockdown library and a set of genes.
}
\section{Slots}{

\describe{
\item{\code{siRNAs}}{The siRNA identifiers}

\item{\code{genes}}{The gene identifiers (Entrez)}

\item{\code{path}}{The path to and \code{.rds} \code{\link{TargetRelations}} file}

\item{\code{loaded}}{An indicator if target relations values are loaded}

\item{\code{values}}{The quantitative target relation values between siRNAs and genes}
}}
\examples{
print("Hello. This is is an example.")
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{TargetRelations}}
}

