% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{TargetRelations}
\alias{TargetRelations}
\alias{TargetRelations,Matrix-method}
\alias{TargetRelations,character-method}
\title{Constructor for TargetRelations objects}
\usage{
TargetRelations(targets)

\S4method{TargetRelations}{character}(targets)

\S4method{TargetRelations}{Matrix}(targets)
}
\arguments{
\item{targets}{Path to a .rds target relations matrix file or matrix object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
This generic function can handle different types of inputs for TargetRelations data. It either takes ...
}
\author{
Fabian Schmich
}

