% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{unloadMatrix}
\alias{unloadMatrix}
\alias{unloadMatrix,TargetRelations-method}
\title{Unload matrix of a TargetRelations object}
\usage{
unloadMatrix(object, ...)

\S4method{unloadMatrix}{TargetRelations}(object, writeMatrix = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Unload matrix of a TargetRelations object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{loadMatrix}}
}

