% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{unloadValues}
\alias{unloadValues}
\alias{unloadValues,TargetRelations-method}
\alias{unloadValues,gespeR-method}
\title{Unload values of a TargetRelations object}
\usage{
unloadValues(object, ...)

\S4method{unloadValues}{TargetRelations}(object, writeValues = TRUE,
  overwrite = FALSE)

\S4method{unloadValues}{gespeR}(object, writeValues = TRUE,
  overwrite = FALSE, path = NULL)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object or \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object or \code{\linkS4class{gespeR}} object
}
\description{
Unload values of a TargetRelations object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{loadValues}}
}

