% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{writeMatrix}
\alias{path<-,TargetRelations,character-method}
\alias{writeMatrix}
\alias{writeMatrix,TargetRelations-method}
\title{Write down the values of a \code{\link{TargetRelations}} object}
\usage{
writeMatrix(object, ...)

\S4method{writeMatrix}{TargetRelations}(object, overwrite = FALSE)

\S4method{path}{TargetRelations,character}(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Write down the values of a \code{\link{TargetRelations}} object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{loadMatrix}} \code{\link{unloadMatrix}}
}

