% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{writeValues}
\alias{writeValues}
\alias{writeValues,TargetRelations-method}
\title{Write down the values of a \code{\link{TargetRelations}} object}
\usage{
writeValues(object, ...)

\S4method{writeValues}{TargetRelations}(object, overwrite = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}

\item{overwrite}{Indicator, whether to overwrite values if file exists at path}

\item{...}{Additional arguments}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Write down the values of a \code{\link{TargetRelations}} object
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{loadValues}} \code{\link{unloadValues}}
}

