% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gespeR-methods.R
\docType{methods}
\name{annotate.gsp}
\alias{annotate.gsp}
\alias{annotate.gsp,Phenotypes-method}
\alias{annotate.gsp,gespeR-method}
\title{annotate.gsp}
\usage{
\S4method{annotate.gsp}{Phenotypes}(object, organism = "hsapiens")

\S4method{annotate.gsp}{gespeR}(object, organism = "hsapiens")
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} or \code{\linkS4class{Phenotypes}} object}

\item{organism}{String indicating the biomaRt organism}
}
\value{
data.frame containing gene identifier, gene symbol and phenotypic score
}
\description{
Query Biomart HGNC symbols for the entrez identifiers of estimated GSPs.
Currently, only implemented for species "hsapiens".
}
\examples{
data(stabilityfits)
 gspA <- gsp(stabilityfits$A)
\dontrun{
 annotate.gsp(gspA)
}
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{gsp}}

\code{\link{ssp}}

\code{\link{scores}}
}

