% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-class.R
\docType{class}
\name{gespeR-class}
\alias{gespeR}
\alias{gespeR,Phenotypes,TargetRelations-method}
\alias{gespeR,numeric,Matrix-method}
\alias{gespeR-class}
\title{gespeR}
\usage{
gespeR(phenotypes, target.relations, ...)

\S4method{gespeR}{Phenotypes,TargetRelations}(phenotypes, target.relations,
  mode = c("cv", "stability"), alpha = 0.5, nbootstrap = 100,
  fraction = 0.67, threshold = 0.9, EV = 1, weakness = 0.8,
  ncores = 1, ...)

\S4method{gespeR}{numeric,Matrix}(phenotypes, target.relations, ...)
}
\arguments{
\item{phenotypes}{The siRNA-spefic phenotypes. Single object for univariate
phenotypes and list of \code{\link{Phenotypes}} objects for multivariate
phenotypes.}

\item{target.relations}{The siRNA-to-gene target relations}

\item{...}{Additional arguments}

\item{mode}{The mode of covariate selectino ("cv" or "stability")}

\item{alpha}{The \code{\link{glmnet}} mixing parameter}

\item{nbootstrap}{The number of bootstrap samples}

\item{fraction}{The fraction for each bootstrap sample}

\item{threshold}{The selection threshold}

\item{EV}{The expected value of wrongly selected elements}

\item{weakness}{The weakness parameter for randomised lasso}

\item{ncores}{The number of cores for parallel computation}
}
\value{
A \code{\linkS4class{gespeR}} object
}
\description{
Class that represents a gespeR model. It contains a SSP \code{\linkS4class{Phenotypes}} and \code{\linkS4class{TargetRelations}}
representing a siRNA knockdown experiment. When the model is fitted, it additionaly contains estimated GSP \code{\linkS4class{Phenotypes}}.
}
\section{Slots}{

\describe{
\item{\code{SSP}}{The observed siRNA-specific phenotypes}

\item{\code{GSP}}{The deconvoluted gene-specific phenotypes}

\item{\code{target.relations}}{The siRNA-to-gene target relations, e.g. predicted by TargetScan}

\item{\code{is.fitted}}{An indicator wheter the gespeR model was fitted}

\item{\code{model}}{The fitted regularized linear regression model}
}}
\examples{
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
trels <- TargetRelations(readRDS(system.file("extdata", "TR_screen_A.rds", package = "gespeR")))
res <- gespeR(phenotypes = phenos,
    target.relations = trels,
    mode = "stability",
    nbootstrap = 100,
    fraction = 0.67,
    threshold = 0.75,
    EV = 1,
    weakness = 0.8,
    ncores = 1)
gsp(res)
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{gespeR-package}}

\code{\link{plot.gespeR}}

\code{\link{gsp}}

\code{\link{ssp}}

\code{\link{scores}}

\code{\link{stability}}

\code{\link{target.relations}}
}

