% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{gespeR-class}
\alias{gespeR-class}
\title{Class definition for gespeR}
\description{
Class that represents a gespeR model. It includes SSP phenotypes and a TargetRelations object
from an siRNA knockdown experiment. When fitted, it additionaly stores the regularized linear regression model
and estimated GSP phenotypes.
}
\section{Slots}{

\describe{
\item{\code{SSP}}{The observed siRNA-specific phenotypes}

\item{\code{GSP}}{The deconvoluted gene-specific phenotypes}

\item{\code{target.relations}}{The siRNA-to-gene target relations, e.g. predicted by TargetScan}

\item{\code{is.fitted}}{An indicator wheter the gespeR model was fitted}

\item{\code{model}}{The fitted regularized linear regression model}
}}
\examples{
print("Hello. This is is an example.")
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{gespeR}}
}

