% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{join}
\alias{join}
\alias{join,TargetRelations,Phenotypes-method}
\title{join}
\usage{
join(targets, phenotypes)

\S4method{join}{TargetRelations,Phenotypes}(targets, phenotypes)
}
\arguments{
\item{targets}{A \code{\linkS4class{TargetRelations}} object.}

\item{phenotypes}{A \code{\linkS4class{Phenotypes}} object.}
}
\value{
List containing the matched targets and phenotypes
}
\description{
Join a TargetRelations object and a Phenotype object
}
\author{
Fabian Schmich

Fabian Schmich
}

