% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Phenotypes-class.R
\docType{methods}
\name{na.rem}
\alias{na.rem}
\alias{na.rem,Phenotypes-method}
\title{Remove NA/Inf values from phenotype vectors}
\usage{
na.rem(object)

\S4method{na.rem}{Phenotypes}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{Phenotypes}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object without NA scores values
}
\description{
Remove NA/Inf values from phenotype vectors
}
\examples{
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
na.rem(phenos)
}
\author{
Fabian Schmich
}

