% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{path<-}
\alias{path<-}
\alias{path<-,TargetRelations,character-method}
\title{path}
\usage{
path(object) <- value

\S4method{path}{TargetRelations,character}(object) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}

\item{value}{A string defining the path}
}
\value{
A \code{\linkS4class{TargetRelations}} object with set path
}
\description{
Set the path of a a \code{\linkS4class{TargetRelations}} object object
}
\examples{
trels <- TargetRelations(readRDS(system.file("extdata", "TR_screen_A.rds", package = "gespeR")))
path(trels) <- "/dev/null"
}
\author{
Fabian Schmich
}

