% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Phenotypes-class}
\alias{Phenotypes-class}
\title{Class definition for Phenotypes}
\description{
Class used to represent various types of phenotypes, e.g. from
siRNA-specific knockdowns (SSP) or estimated gene-specific phenotypes (GSP).
}
\section{Slots}{

\describe{
\item{\code{type}}{The indicator of represented phenotypes (i.e., "SSP" or "GSP")}

\item{\code{ids}}{The phenotype identifiers (i.e., siRNA or gene ids)}

\item{\code{values}}{The phenotypic values}
}}
\examples{
print("Hello. This is is an example.")
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{Phenotypes}}
}

