% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-class.R
\name{plot.gespeR}
\alias{plot.gespeR}
\title{Plot method for \code{\linkS4class{gespeR}} objects}
\usage{
\method{plot}{gespeR}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{gespeR}} object}

\item{...}{Additional paramters for plot}
}
\value{
Histogram of SSPs or GSPs
}
\description{
Plot method for \code{\linkS4class{gespeR}} objects
}
\author{
Fabian Schmich
}

