% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Phenotypes-class.R
\name{plot.Phenotypes}
\alias{plot.Phenotypes}
\title{Plot method for Phenotype objects}
\usage{
\method{plot}{Phenotypes}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{...}{Additional arguments for plot}
}
\value{
Histogram of scores
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
plot(phenos)
}
\description{
Plot method for Phenotype objects
}
\author{
Fabian Schmich
}

