% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.Phenotypes}
\alias{plot.Phenotypes}
\title{Plot method for Phenotype objects}
\usage{
\method{plot}{Phenotypes}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{...}{Additional arguments for plot}
}
\value{
Histogram of scores
}
\description{
Plot method for Phenotype objects
}
\author{
Fabian Schmich
}

