% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbo}
\alias{rbo}
\title{Rank biased overlap (Webber et al., 2010)}
\usage{
rbo(s, t, p, k = floor(max(length(s), length(t))/2), side = c("top",
  "bottom"), mid = NULL)
}
\arguments{
\item{s}{List 1}

\item{t}{List 2}

\item{p}{Weighting parameter in [0, 1]. High p implies strong emphasis on top ranked elements}

\item{k}{Evaluation depth for extrapolation}

\item{side}{Evaluate similarity between the top or the bottom of the ranked lists}

\item{mid}{Set the mid point to for example only consider positive or negative scores}
}
\value{
rank biased overlap (rbo)
}
\description{
Evaluates the rank biased overlap (rbo) of two ranked lists based on formula based on (32) from
"A Similarity Measure for Indefinite Rankings" (Webber et al.). Two ranked lists with high rbo are
very similar, wheras low rbo indicates dissimilar lists. rbo ranges between 0 and 1. In this method
the extrapolated version of rbo is implemented.
}
\examples{
a <- rnorm(26)
b <- rnorm(26)
names(a) <- names(b) <- LETTERS
rbo(a, b, p = 0.95)
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{concordance}}
}

