% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-package.R
\docType{data}
\name{simData}
\alias{simData}
\title{Example data: Simulated phenotypes and target relations for 4 screens (A, B, C, D)}
\description{
The data set contains simulated data for four screens. Each screen consists of a phenotype vector
and target relations between siRNAs and genes, i.e. which siRNA binds which genes (on- and off-targets).
The size of each simulated screen is N = 1000 siRNAs x p = 1500 genes.
SSPs are generated by first defining GSPs and multiplying the true GSPs with the sampled targe trelation
matrices. For sampling the GSPs, we set the number of effect genes to 5% and its effect strengths are sampled
from Normal(0, 3).
Target relation matrices are simulated by sampling the number of off-targets per siRNA from
Normal(3e-2 * N, 3e-3 * N) and the strength of off-targets is sampled from Beta(2, 5). On-target
components are set to 0.75.
}
\details{
The code used to simluate the data can be found in system.file("example", "data_simulation.R", package="gespeR")
}
\examples{
pheno.a <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package="gespeR"),
type = "SSP", col.id = 1, col.score = 2)
targets.a <- TargetRelations(system.file("extdata", "TR_screen_A.rds", package="gespeR"))
}

