% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-class.R
\docType{methods}
\name{stability}
\alias{stability}
\alias{stability,gespeR-method}
\title{stability}
\usage{
stability(object)

\S4method{stability}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object of SSPs
}
\description{
Retrieve a \code{\linkS4class{Phenotypes}} object with stability values from a
\code{\linkS4class{gespeR}} object.
}
\examples{
phenos <- Phenotypes(system.file("extdata", "Phenotypes_screen_A.txt", package = "gespeR"),
type = "SSP",
col.id = 1,
col.score = 2)
trels <- TargetRelations(readRDS(system.file("extdata", "TR_screen_A.rds", package = "gespeR")))
res <- gespeR(phenotypes = phenos,
 target.relations = trels,
 mode = "stability",
 nbootstrap = 100,
 fraction = 0.67,
 threshold = 0.75,
 EV = 1,
 weakness = 0.8,
 ncores = 1)
stab <- stability(res)
ans <- merge(as.data.frame(gsp(res)), as.data.frame(stability(res)), by = "ID")
colnames(ans)[2:3] <- c("Phenotype", "Stability")
ans[order(ans$Stability, decreasing = TRUE),]
}
\author{
Fabian Schmich
}

