% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{stability}
\alias{stability}
\alias{stability,gespeR-method}
\title{stability}
\usage{
stability(object)

\S4method{stability}{gespeR}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{gespeR}} object}
}
\value{
A \code{\linkS4class{Phenotypes}} object of SSPs
}
\description{
Retrieve a \code{\linkS4class{Phenotypes}} object with stability values from a
\code{\linkS4class{gespeR}} object.
}
\author{
Fabian Schmich
}

