% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-package.R
\docType{data}
\name{stabilityfits}
\alias{stabilityfits}
\title{Example fits for phenotypes from example screening data A, B, C and D}
\description{
The data set contains four fitted gespeR models using stability selection
}
\examples{
data(stabilityfits)
}

