% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,Phenotypes,ANY,ANY,ANY-method}
\alias{[,Phenotypes,ANY,ANY,ANY-method}
\title{Subsetting for Phenotype objects.}
\usage{
\S4method{[}{Phenotypes,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{i}{The subsetting indices}

\item{j}{Subsetting indices (not used)}

\item{...}{Additional parameters}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\value{
A \code{\linkS4class{Phenotypes}} object
}
\description{
Subsetting for Phenotype objects.
}
\author{
Fabian Schmich
}

