% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gespeR-generics.R
\docType{methods}
\name{[,Phenotypes,ANY,ANY,ANY-method}
\alias{[,Phenotypes,ANY,ANY,ANY-method}
\title{Subsetting for Phenotype objects.}
\usage{
\S4method{[}{Phenotypes,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{Phenotypes}} object}

\item{i}{The subsetting indices}

\item{j}{Subsetting indices (not used)}

\item{...}{Additional parameters}

\item{drop}{Drop Redundant Extent Information}
}
\value{
A \code{\linkS4class{Phenotypes}} object
}
\description{
Subsetting for Phenotype objects.
}
\author{
Fabian Schmich
}

