% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gespeR-generics.R
\docType{methods}
\name{[,TargetRelations,ANY,ANY,ANY-method}
\alias{[,TargetRelations,ANY,ANY,ANY-method}
\title{Subsetting for TargetRelations objects.}
\usage{
\S4method{[}{TargetRelations,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{TargetRelations}} object}

\item{i}{The row subsetting indices (siRNAs)}

\item{j}{The column subsetting indeces (genes)}

\item{...}{Additional parameters}

\item{drop}{Drop Redundant Extent Information}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Subsetting for TargetRelations objects.
}
\author{
Fabian Schmich
}

