% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,TargetRelations,ANY,ANY,ANY-method}
\alias{[,TargetRelations,ANY,ANY,ANY-method}
\title{Subsetting for TargetRelations objects.}
\usage{
\S4method{[}{TargetRelations,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{TargetRelations}} object}

\item{i}{The row subsetting indices (siRNAs)}

\item{j}{The column subsetting indeces (genes)}

\item{...}{Additional parameters}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Subsetting for TargetRelations objects.
}
\author{
Fabian Schmich
}

