% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TargetRelations-class}
\alias{TargetRelations}
\alias{TargetRelations,Matrix-method}
\alias{TargetRelations,character-method}
\alias{TargetRelations-class}
\title{TargetRelations}
\usage{
TargetRelations(targets)

\S4method{TargetRelations}{character}(targets)

\S4method{TargetRelations}{Matrix}(targets)
}
\arguments{
\item{targets}{Path to a .rds target relations matrix file or \code{\link{Matrix}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Class used to represent siRNA-to-gene on- and off-target relations
for a knockdown library and a set of genes.
}
\section{Slots}{

\describe{
\item{\code{siRNAs}}{The siRNA identifiers}

\item{\code{genes}}{The gene identifiers (Entrez)}

\item{\code{path}}{The path to and \code{.rds} \code{\linkS4class{TargetRelations}} file}

\item{\code{is.loaded}}{An indicator if target relations values are loaded}

\item{\code{values}}{The quantitative target relation values between siRNAs and genes}
}}
\examples{
trels <- TargetRelations(readRDS(system.file("extdata", "TR_screen_A.rds", package = "gespeR")))
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{join}}

\code{\link{loadValues}}

\code{\link{unloadValues}}

\code{\link{writeValues}}

\code{\link{values}}

\code{\link{path<-}}
}

