% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TargetRelations-class}
\alias{TargetRelations}
\alias{TargetRelations,Matrix-method}
\alias{TargetRelations,character-method}
\alias{TargetRelations-class}
\title{TargetRelations}
\usage{
TargetRelations(targets)

\S4method{TargetRelations}{character}(targets)

\S4method{TargetRelations}{Matrix}(targets)
}
\arguments{
\item{targets}{Path to a .rds target relations matrix file or \code{\link{Matrix}} object}
}
\value{
A \code{\linkS4class{TargetRelations}} object
}
\description{
Class used to represent siRNA-to-gene on- and off-target relations
for a knockdown library and a set of genes.
}
\section{Slots}{

\describe{
\item{\code{siRNAs}}{The siRNA identifiers}

\item{\code{genes}}{The gene identifiers (Entrez)}

\item{\code{path}}{The path to and \code{.rds} \code{\linkS4class{TargetRelations}} file}

\item{\code{is.loaded}}{An indicator if target relations values are loaded}

\item{\code{values}}{The quantitative target relation values between siRNAs and genes}
}}
\examples{
# Read phenotypes
phenos <- lapply(LETTERS[1:4], function(x) {
  sprintf("Phenotypes_screen_\%s.txt", x)
  })
phenos <- lapply(phenos, function(x) {
  Phenotypes(system.file("extdata", x, package="gespeR"),
             type = "SSP",
             col.id = 1,
             col.score = 2)
})  
phenos
plot(phenos[[1]])

# Read target relations
tr <- lapply(LETTERS[1:4], function(x) {
  sprintf("TR_screen_\%s.rds", x)
})
tr <- lapply(tr, function(x) {
  TargetRelations(system.file("extdata", x, package="gespeR"))
})
tr[[1]]
tempfile <- paste(tempfile(pattern = "file", tmpdir = tempdir()), ".rds", sep="")
tr[[1]] <- unloadValues(tr[[1]], writeValues = TRUE, path = tempfile)
tr[[1]]
tr[[1]] <- loadValues(tr[[1]])
tr[[1]]

# Fit gespeR models with cross validation
res.cv <- lapply(1:length(phenos), function(i) {
  gespeR(phenotypes = phenos[[i]],
         target.relations = tr[[i]],
         mode = "cv",
         alpha = 0.5,
         ncores = 1)
})
summary(res.cv[[1]])
res.cv[[1]]
plot(res.cv[[1]])

# Extract scores
ssp(res.cv[[1]])
gsp(res.cv[[1]])
head(scores(res.cv[[1]]))

# Fit gespeR models with stability selection
res.stab <- lapply(1:length(phenos), function(i) {
  gespeR(phenotypes = phenos[[i]],
         target.relations = tr[[i]],
         mode = "stability",
         nbootstrap = 100,
         fraction = 0.67,
         threshold = 0.75,
         EV = 1,
         weakness = 0.8,
         ncores = 1)
})
summary(res.stab[[1]])
res.stab[[1]]
plot(res.stab[[1]])

# Extract scores
ssp(res.stab[[1]])
gsp(res.stab[[1]])
head(scores(res.stab[[1]]))

# Compare concordance between stability selected GSPs and SSPs
conc.gsp <- concordance(lapply(res.stab, gsp))
conc.ssp <- concordance(lapply(res.stab, ssp))

pl.gsp <- plot(conc.gsp) + ggtitle("GSPs\\n")
pl.ssp <- plot(conc.ssp) + ggtitle("SSPs\\n")

if (require(grid)) {
  grid.newpage()
  pushViewport(viewport(layout = grid.layout(1, 2) ) )
  print(pl.gsp, vp = viewport(layout.pos.row = 1, layout.pos.col = 1))
  print(pl.ssp, vp = viewport(layout.pos.row = 1, layout.pos.col = 2))
} else {
  plot(pl.gsp)
  plot(pl.ssp)
}
}
\author{
Fabian Schmich
}
\seealso{
\code{\link{join}}

\code{\link{loadValues}}

\code{\link{unloadValues}}

\code{\link{writeValues}}

\code{\link{values}}

\code{\link{path<-}}
}

