% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{values}
\alias{values}
\alias{values,TargetRelations-method}
\title{values}
\usage{
values(object)

\S4method{values}{TargetRelations}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\link{Matrix}} object
}
\description{
Retrieve the numeric siRNA-to-gene target relations from a \code{\linkS4class{TargetRelations}} object
}
\examples{
trels <- TargetRelations(readRDS(system.file("extdata", "TR_screen_A.rds", package = "gespeR")))
values(trels)[1:5, 1:5]
}
\author{
Fabian Schmich
}

