% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{values}
\alias{values}
\alias{values,TargetRelations-method}
\title{Get values}
\usage{
values(object)

\S4method{values}{TargetRelations}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TargetRelations}} object}
}
\value{
A \code{\linkS4class{Matrix}} object of TargetRelations values
}
\description{
Get values
}
\author{
Fabian Schmich
}

